DAT-4115 old side nav cleanup
