--- front/main/app/templates/components/side-nav-local-navigation-menu.hbs
@@ -1,41 +1,0 @@
-{{#if localNavContent.parent}}
-	<li class="side-nav-menu__item main">
-		{{localNavContent.text}}
-	</li>
-{{else}}
-	<li class="side-nav-menu__item main">
-		{{i18n 'app.side-nav-menu-label'}}
-	</li>
-{{/if}}
-
-{{#unless localNavContent.parent}}
-	<li class="side-nav-menu__item" {{action 'recentWikiActivityClick'}}>
-		{{#link-to 'recent-wiki-activity'
-			trackingCategory='recent-wiki-activity'
-			trackingLabel='local-nav'
-		}}
-			{{i18n 'main.title' ns='recent-wiki-activity'}}
-			{{#if shouldDisplayNewBadge}}
-				{{svg 'info-filled' class='icon new-badge'}}
-			{{/if}}
-		{{/link-to}}
-	</li>
-{{/unless}}
-
-{{#each localNavContent.children as |child index|}}
-	{{#if child.children}}
-		<li class="side-nav-menu__item menu" {{action 'changeMenuItem' index}}>
-			{{unbound child.text}}
-		</li>
-	{{else}}
-		<li class="side-nav-menu__item mw-content">
-			<a {{action 'clickLink' index preventDefault=false}} href="{{unbound child.href}}">{{unbound child.text}}</a>
-		</li>
-	{{/if}}
-{{/each}}
-
-{{#unless localNavContent.parent}}
-	<li class="side-nav-menu__item" {{action 'loadRandomArticle'}}>
-		{{i18n 'app.random-page-label'}}
-	</li>
-{{/unless}}
