--- front/main/app/templates/components/side-nav-local-navigation-root.hbs
@@ -1,14 +1,0 @@
-{{#if localNavigationVisible}}
-	<div class="side-nav-menu__item back" {{action 'goBack' backTarget}}>
-		{{i18n 'app.navigation-back-label'}}
-	</div>
-{{/if}}
-
-{{#if localNavigationVisible}}
-	{{side-nav-local-navigation-menu
-	collapse=attrs.collapse
-	loadRandomArticle=this.attrs.loadRandomArticle
-	localNavContent=localNavContent
-	updateContent=(action 'updateContent')
-	}}
-{{/if}}
